A single unrelated calculation that does not comply with standards causes compliance to fail for all other calculations (see failing test)
This is a problem because our emitter models now have committees for multiple impacts, but standards are usually specific to a single impact. For example the GHG Protocol only applies to carbon impact, so quorums in committees for unrelated impacts (e.g. energy use, water use, waste production) will not comply with the GHG Protocol. But compliance is evaluated across all quorums, regardless of which impact that quorum's committee relates to. This means that as soon as you add a committee for a non-carbon impact, no calculations comply with any standards.